move directory parsing code to modules
